Move database funcs into helper
